requirements.txt
pytorch-lightning==2.1.2
gradio==3.41.2
huggingface-hub
einops
omegaconf
torchmetrics
webdataset
accelerate
tensorboard
PyMCubes
trimesh
rembg
transformers==4.34.1
diffusers==0.20.2
bitsandbytes
imageio[ffmpeg]
xatlas
plyfile
wandb
albumentations
mmcv-full==1.7.2
kiui >= 0.2.3
pymeshlab
deepspeed
tyro
nerfacc
git+https://github.com/ashawkey/diff-gaussian-rasterization.git
git+https://github.com/Stability-AI/generative-models.git
